FINALY

+++ num_vector_utils.h
@@ -8,6 +8,6 @@ int num_vector_product(const vector<int>&);
 vector<int> only_evens(const vector<int>&);
 vector<int> only_odds(const vector<int>&);
 vector<int> nums_between(const vector<int>&,int low,int high);
-double num_vector_mean(const vector<int>&);
+double num_vector_mean(const vector<double>&);
 /*int num_vector_median(cosnt vector<int>&);
 int num_vector_mode(cosnt vector<int>&);*/
